reorganized imports and moved enums to the file that is closely assosiated with the variable.
