+++ lib/features/file_parsing/file_parsing_barrel.dart
@@ -1,3 +1,4 @@
 export 'edit_file_name.dart';
 export 'file_input.dart';
 export 'file_output.dart';
+export 'file_parser.dart';
