+++ lib/features/media_conversion/media_conversion_barrel.dart
@@ -1,2 +1,3 @@
+export 'conversion_cmd.dart';
 export 'convert_button.dart';
 export 'media.dart';
